requirements.txt
rx
pyside2
asyncqt
bottom
sortedcontainers
aiohttp
docopt
pyyaml
